+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `ConstantsExtension`
+- `PhpExtension`
 
 ## [1.2.3](https://github.com/orisai/nette-di/compare/1.2.2...1.2.3) - 2022-12-14