+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-di/compare/1.3.0...HEAD)
 
+## [1.3.1](https://github.com/orisai/nette-di/compare/1.3.0...1.3.1) - 2023-03-22
+
 ### Fixed
 
 - `BaseConfigurator`